[change] Port ScrollView to Class Component
Port ScrollView from legacy createReactClass to ES6 Class syntax.
