+++ packages/benchmarks/package.json
@@ -13,11 +13,11 @@
     "prop-types": "^15.6.0",
     "react": ">=17.0.2",
     "react-dom": ">=17.0.2",
-    "react-native-web": "0.18.9"
+    "react-native-web": "0.18.10"
   },
   "devDependencies": {
     "babel-loader": "^8.2.5",
-    "babel-plugin-react-native-web": "0.18.9",
+    "babel-plugin-react-native-web": "0.18.10",
     "css-loader": "^6.7.1",
     "style-loader": "^3.3.1",
     "url-loader": "^4.1.1",
