+++ packages/react-native-web/package.json
@@ -23,7 +23,6 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.18.6",
-    "create-react-class": "^15.7.0",
     "fbjs": "^3.0.4",
     "inline-style-prefixer": "^6.0.1",
     "normalize-css-color": "^1.0.2",
